analyzer/requirements.txt
python-chess
pymongo
tqdm
pandas
numpy